requirements.txt
streamlit>=1.30.0
openai>=1.12.0
tavily-python>=0.5.0
neo4j>=5.15.0
python-dotenv>=1.0.0
pyvis>=0.3.2
plotly>=5.18.0
